Clean up and consolidate UI layout
Remove some unnecessary UI elements. Fix overflow in tree view. Tweak
styling so that tabs can be merged into the panel header (and are
visually consistent with static headers). Fix layout of top header,
which was not size-responsive (not the only issue we have in that
department, but it's an improvement). Tweak assorted other elements for
improved aesthetics.