Add frame number display
Add an area that shows (and allows changing) the current frame number
next to the scrubber. This is supposed to be a spin box, but so far I
haven't figured out a good way to reserve an appropriate amount of space
(I am REALLY trying to resist a hard-coded pixel value), and the up/down
buttons don't feel like a critical-importance item for now, so probably
won't be implemented until some time in the future. (Using a native spin
box is really ugly, since those don't support added text, resulting in
the buttons being in the middle of text. Ick. Also, at least in Firefox,
they like to overflow the frame for some unknown reason.)