Fix frame range
Change the 'empty' frame range back to [MAX, MIN] (albeit using integers
this time), since this is needed for expanding an empty range to work
properly. This reportedly causes issues on the front-end, but having our
"empty" state be distinguishable from a valid state is probably a Good
Thing, and so we should probably fix the front-end to cope. (Immediately
prior to this change, the frame range would get "stuck" always having 0
as the lower bound.)