Implement low-level activity [track] removal
Add methods to AnnotationActivityContainer to remove an activity, or to
remove a track (from all activities that use the track). The purpose of
the first is obvious. The purpose of the latter is to clean up the
activities when a track is deleted. Note that, if removing a track
causes an activity to have an empty track set, the activity itself will
be deleted as well.